Part6.a: Redux notes
Details:
---
* Separate the logic into the `<Notes />` *container component*,
  changing the `<App />` and the new `<Note />` into purely
  presentational components. The `<Notes />` component is the only one
  accessing the Redux store.
