+++ part6/redux-notes/src/Note.jsx
@@ -1,0 +1,7 @@
+const Note = ({ note, handleClick }) => (
+    <li onClick={handleClick}>
+        {note.content} <strong>{note.important ? 'important' : ''}</strong>
+    </li>
+)
+
+export default Note
